+++ packages/fighting-design/_hooks/use-input/index.ts
@@ -1,7 +1,7 @@
-import { isNumber } from '../../_utils'
 import { useRun } from '..'
 import { EMIT_UPDATE } from '../../_tokens'
 import type { InputType } from '../../input'
+import type { WritableComputedRef } from 'vue'
 
 /**
  * 传入的 props 类型接口
@@ -12,23 +12,23 @@ import type { InputType } from '../../input'
  * @param { string } [type] 文本框类型
  */
 export interface UseInputProps {
-  onInput?: (val: string) => void | null
-  onChange?: (val: string) => void | null
-  disabled?: boolean
-  type?: InputType
+  onInput: Function
+  onChange: Function
+  disabled: boolean
+  type: InputType
 }
 
 /**
  * useInput hook 返回值类型接口
  *
- * @param { Function } onInput 处理文本框输入
- * @param { Function } onChange 改变后触发的回调
- * @param { Function } onClear 清空文本框
+ * @param { Function } handleInput 处理文本框输入
+ * @param { Function } handleChange 改变后触发的回调
+ * @param { Function } handleClear 清空文本框
  */
 export interface UseInputReturn {
-  onInput: (evt: Event) => void
-  onChange: (evt: Event) => void
-  onClear: () => void
+  handleInput: (evt: Event) => void
+  handleChange: (evt: Event) => void
+  handleClear: () => void
 }
 
 /**
@@ -49,7 +49,8 @@ export type UseInputEmit = (event: 'update:modelValue', val: string | number) =>
  * @param { Function } emit 回调参数
  * @returns { Object }
  */
-export const useInput = (prop: UseInputProps, emit: UseInputEmit): UseInputReturn => {
+export const useInput = (prop: Partial<UseInputProps>, emit: UseInputEmit, keyword: WritableComputedRef<string | number>): UseInputReturn => {
+
   const { run } = useRun()
 
   /**
@@ -57,15 +58,8 @@ export const useInput = (prop: UseInputProps, emit: UseInputEmit): UseInputRetur
    *
    * @param { Object } evt 事件对象
    */
-  const onInput = (evt: Event): void => {
-    emit(
-      EMIT_UPDATE,
-      isNumber(prop.type)
-        ? Number((evt.target as HTMLInputElement).value)
-        : (evt.target as HTMLInputElement).value
-    )
-
-    run(prop.onInput, (evt.target as HTMLInputElement).value, evt)
+  const handleInput = (evt: Event): void => {
+    run(prop.onInput, keyword.value, evt)
   }
 
   /**
@@ -73,19 +67,19 @@ export const useInput = (prop: UseInputProps, emit: UseInputEmit): UseInputRetur
    *
    * @param { Object } evt 事件对象
    */
-  const onChange = (evt: Event): void => {
-    run(prop.onChange, (evt.target as HTMLInputElement).value, evt)
+  const handleChange = (evt: Event): void => {
+    run(prop.onChange, keyword.value, evt)
   }
 
   /** 清空文本框 */
-  const onClear = (): void => {
+  const handleClear = (): void => {
     if (prop.disabled) return
     emit(EMIT_UPDATE, '')
   }
 
   return {
-    onInput,
-    onChange,
-    onClear
+    handleInput,
+    handleChange,
+    handleClear
   }
 }