+++ packages/fighting-design/textarea/src/textarea.vue
@@ -1,36 +1,18 @@
 <script lang="ts" setup name="FTextarea">
   import { Props } from './props'
-  import { useInput, useProps, useList } from '../../_hooks'
-  import { isString, isNumber } from '../../_utils'
+  import { useInput, useList, useModel } from '../../_hooks'
   import { FIconCross } from '../../_svg'
   import { FSvgIcon } from '../../svg-icon'
   import { EMIT_UPDATE } from '../../_tokens'
 
   const prop = defineProps(Props)
   const emit = defineEmits({
-    [EMIT_UPDATE]: (val: string | number): boolean => isString(val) || isNumber(val)
+    [EMIT_UPDATE]: (val: string | number): boolean => !!val
   })
 
-  const { filter } = useProps(prop)
-  const { onInput, onClear, onChange } = useInput(
-    filter(['onChange', 'onInput', 'disabled']),
-    emit
-  )
+  const { keyword } = useModel(prop, 'modelValue', emit, EMIT_UPDATE)
+  const { handleInput, handleClear, handleChange } = useInput(prop, emit, keyword)
   const { classes, styles } = useList(prop, 'textarea')
-
-  /**
-   * 输入框输入
-   *
-   * @param { Object } evt 事件对象
-   */
-  const handleInput = (evt: Event): void => onInput(evt)
-
-  /**
-   * 文本输入 change 事件
-   *
-   * @param { Object } evt 事件对象
-   */
-  const handleChange = (evt: Event): void => onChange(evt)
 
   /** 类名列表 */
   const classList = classes(['disabled'], 'f-textarea')
@@ -42,18 +24,18 @@
 <template>
   <div :class="classList" :style="styleList">
     <textarea
+      v-model="keyword"
       class="f-textarea__textarea"
       :rows="rows"
-      :value="modelValue"
       :disabled="disabled"
       :readonly="readonly"
       :autofocus="autofocus"
       :placeholder="placeholder"
       :name="name"
-      @input="handleInput"
-      @blur="onBlur"
-      @focus="onFocus"
-      @change="handleChange"
+      @input="handleInput($event)"
+      @change="handleChange($event)"
+      @blur="onBlur($event)"
+      @focus="onFocus($event)"
     />
 
     <!-- 清空按钮 -->
@@ -62,7 +44,7 @@
       class="f-textarea__clear-btn"
       :icon="FIconCross"
       :size="14"
-      @click="onClear"
+      @click="handleClear"
     />
   </div>
 </template>